refactor: extract admin helpers and clean up docker-compose
Extract getLast7Days and fillZeroDays into adminHelpers.ts.
Remove commented-out certbot port block from docker-compose.production.yml.
