+++ apps/backend/src/api/routes/admin.route.ts
@@ -1,24 +1,7 @@
 import { FastifyPluginAsync } from 'fastify'
 import { sendError } from '../helpers'
 import { prisma } from '@/lib/prisma'
-
-function getLast7Days(): string[] {
-  const dates: string[] = []
-  const now = new Date()
-  for (let i = 6; i >= 0; i--) {
-    const d = new Date(now.getTime() - i * 24 * 60 * 60 * 1000)
-    dates.push(d.toISOString().slice(0, 10))
-  }
-  return dates
-}
-
-function fillZeroDays(
-  rows: { date: string; count: bigint }[],
-  days: string[]
-): { date: string; count: number }[] {
-  const map = new Map(rows.map((r) => [r.date, Number(r.count)]))
-  return days.map((date) => ({ date, count: map.get(date) ?? 0 }))
-}
+import { getLast7Days, fillZeroDays } from '../adminHelpers'
 
 const adminRoutes: FastifyPluginAsync = async (fastify) => {
   // Defense in depth: require trusted header set by nginx mTLS admin proxy
